Put all the 'tools' in a separate folder
