+++ plotComparedDistriFromTwoFiles.py
@@ -1,5 +1,5 @@
 from ROOT import TFile, gROOT, TH1D, kRed, TLegend
-from drawCanvas import *
+from tools.drawCanvas import *
 import os
 
 gROOT.SetBatch()
