+++ tools/crabOutAnalyzer.py
